Change: Overlay is now split from the original sheet, so it can be added & styled seperately
