+++ src/lib/types.ts
@@ -3,6 +3,7 @@ import type Sheet from './BottomSheet/Sheet/Sheet.svelte';
 import type BottomSheet from './BottomSheet/BottomSheet.svelte';
 import type SheetTrigger from './BottomSheet/Trigger/Trigger.svelte';
 import { writable, type Writable } from 'svelte/store';
+import type Overlay from './BottomSheet/Overlay/Overlay.svelte';
 
 export type BottomSheetSettings = {
 	disableScrollingOutside?: boolean;
@@ -18,6 +19,7 @@ export type BottomSheetType = typeof BottomSheet & {
 	Content: typeof BottomSheetContent;
 	Sheet: typeof Sheet;
 	Trigger: typeof SheetTrigger;
+	Overlay: typeof Overlay;
 };
 
 export type SheetContext = {
